feat(uplink+downlink): expand max file ring buffer size

+++ src/api/downlink.c
@@ -26,7 +26,7 @@ const downlink_row_t downlink_row = {
 		.size = 291,
 };
 
-uint16_t downlink_select_one(octet_t *db, downlink_t *downlink, uint8_t *head) {
+uint16_t downlink_select_one(octet_t *db, downlink_t *downlink, uint16_t *head) {
 	uint16_t status;
 
 	char file[128];
@@ -71,7 +71,7 @@ cleanup:
 	return status;
 }
 
-uint16_t downlink_insert(octet_t *db, downlink_t *downlink, uint8_t *tail) {
+uint16_t downlink_insert(octet_t *db, downlink_t *downlink, uint16_t *tail) {
 	uint16_t status;
 
 	char file[128];
@@ -116,7 +116,7 @@ cleanup:
 	return status;
 }
 
-uint16_t downlink_delete(octet_t *db, downlink_t *downlink, uint8_t *head) {
+uint16_t downlink_delete(octet_t *db, downlink_t *downlink, uint16_t *head) {
 	uint16_t status;
 
 	char file[128];
