+++ src/api/uplink.h
@@ -48,6 +48,6 @@ extern const char *uplink_file;
 
 extern const uplink_row_t uplink_row;
 
-uint16_t uplink_select_one(octet_t *db, uplink_t *uplink, uint8_t *head);
-uint16_t uplink_insert(octet_t *db, uplink_t *uplink, uint8_t *tail);
-uint16_t uplink_delete(octet_t *db, uplink_t *uplink, uint8_t *head);
+uint16_t uplink_select_one(octet_t *db, uplink_t *uplink, uint16_t *head);
+uint16_t uplink_insert(octet_t *db, uplink_t *uplink, uint16_t *tail);
+uint16_t uplink_delete(octet_t *db, uplink_t *uplink, uint16_t *head);
